+++ io-mokamint-node-integration/src/test/java/io/mokamint/node/integration/tests/RemoteRestrictedNodeTests.java
@@ -69,7 +69,7 @@ public class RemoteRestrictedNodeTests extends AbstractLoggedTests {
 		}
 	}
 
-	private final static long TIME_OUT = 500L;
+	private final static int TIME_OUT = 500;
 
 	/**
 	 * Test server implementation.
